[modify] ranking 화면 구현 및 데이터 완료, 클릭 이벤트 필요

+++ runnershi_fe/RunnersHi/app/src/main/java/com/A306/runnershi/Model/Ranking.kt
@@ -3,5 +3,5 @@ package com.A306.runnershi.Model
 data class Ranking (
     var userId: String? = null,
     var userName: String? = null,
-    var userData: Long = 0L
+    var userData: Double = 0.0
 )
